# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /user/CreateUser:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加用户
      tags:
      - 用户模块
  /user/DeleteUser:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/FindUserByNameAndPwd:
    post:
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: password
        in: query
        name: password
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 用户登录
      tags:
      - 用户模块
  /user/GetUserList:
    get:
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 查询所有用户
      tags:
      - 用户模块
  /user/UpdateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
swagger: "2.0"